fixed bug with package rx duration: is now uint32 instead of uint16 start/stop time stamps

+++ src/avr-common/utils/uc-core/synchronization/Deviation.h
@@ -82,7 +82,7 @@ void calculateVarianceAndStdDeviance(TimeSynchronization *const timeSynchronizat
     timeSynchronization->variance /= timeSynchronization->timeIntervalSamples.numSamples;
 
 #ifdef DEVIATION_MATH_SQRT
-    timeSynchronization->stdDeviance = sqrt(timeSynchronization->variance);
+    timeSynchronization->stdDeviance = sqrtf(timeSynchronization->variance);
 #else
 #  if defined(DEVIATION_BINARY_SEARCH_SQRT)
     __binarySearchSqr(&timeSynchronization->variance, &timeSynchronization->stdDeviance);
